requirements.txt
llama-index
llama-index-vector-stores-mongodb
llama-index-embeddings-openai
pymongo
datasets
pandas
trulens_eval
python-dotenv
streamlit
watchdog
IPython
streamlit
pytube
pydub
google-cloud-speech
google-cloud-aiplatform
cohere
weaviate-client
openai
stability-sdk